\name{predict.gcm}
\alias{predict.gcm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function for generating GCM model predictions
}
\description{
This function generates model predictions for an object of class \code{gcm}, typically based on posterior samples generated by the function \link{gcm}.
}
\usage{
predict.gcm(model, user.samples = NULL, newdata = NULL, weights = NULL,
            type = c("response", "probabilities"), 
            sample.type = c("random", "user", "mean"), 
            iterations = 1000)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an object of class \code{gcm} generated by the function \link{gcm}.
}
  \item{user.samples}{
an object of class \code{mcmc}, \code{matrix}, \code{data.frame}, a numeric vector, or \code{NULL}. Allows the user to specify one or more sets of parameter values that are to be used for prediction. The object has to have the same column names (or names, in the case of vectors) as those specified in \code{object$posterior.samples}. If \code{NULL}, the predictions are based on random samples or the mean of the parameter values in \code{object$posterior.samples}, depending on the value of \code{sample.type} (see below).
}
  \item{newdata}{
a \code{data.frame} object containing new items for which predictions are to be made. It is recommended that this \code{data.frame} have the same shape as the input to the \code{gcm} call that generated \code{object}.
}
  \item{weights}{
only relevant when \code{newdata != NULL}. A numeric vector containing the frequencies of the items in \code{newdata}.
}
  \item{type}{
an object of class character. If \code{"response"}, the function returns predicted category labels for each item. If \code{"probabilities"}, the function returns categorisation probabilities for each item. This also affects the \code{summary} and \code{plot} functions that can be run on the output of \code{predict.gcm}: if \code{"response"}, \code{summary} and \code{plot} are based on accuracy per posterior sample; if \code{"probabilities"}, \code{summary} and \code{plot} are based on log-likelihood per posterior sample.
}
  \item{sample.type}{
an object of class character. If \code{"random"}, \code{predict.gcm} generates predictions based on \code{n} random samples (\code{n} = \code{iterations}; see below) from \code{object$posterior.samples} or \code{user.samples} if the latter is not \code{NULL}. If \code{"mean"}, \code{predict.gcm} generates a single set of predictions based on the mean of each parameter in \code{object$posterior.samples} or \code{user.samples} if the latter is not \code{NULL}. If \code{"user"}, \code{predict.gcm} generates predictions based all the samples in \code{user.samples}.
}
  \item{iterations}{
an integer specifying the number of samples to be drawn from \code{object$posterior.samples} or \code{user.samples}.
}
}
\details{
Unless \code{sample.type} is set to \code{"mean"}, \code{predict.gcm} generates predictions for the entire input data set (or \code{newdata}) \code{iterations} number of times, each time based on a different posterior sample. This is based on Kruschke (2010). Each of these is stored separately in the output of \code{predict.gcm}. The functions \code{plot} and \code{summary} can be used to summarise the output.
}
\value{
  \item{data.type}{\code{"response"} or \code{"probabilities"}}
  \item{predictions}{a list with \code{iterations} elements, each of them representing a set of predictions for the input data set based on a different posterior sample. Each element is a list with the the following internal structure:
  \itemize{
    \item \code{index}: the index of the posterior sample in \code{do.call(rbind, object$posterior.sample)} or \code{user.samples}.
    \item \code{parameters}: the parameter estimates from the posterior sample.
    \item \code{outcome}: a vector of predicted category labels if \code{type=="response"}; a matrix containing categorisation probabilities for each outcome category (columns) within each item (rows).
    }
  }
  \item{correct}{a vector of original correct category labels from the data.}
  \item{formula}{the \code{gcm} model formula.}
  }
%% ...
}
\references{
Kruschke, J. (2010) Doing Bayesian data analysis: A tutorial introduction with R. Academic Press.
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
# generating posterior samples

model <- gcm(category~size + colour + shape + symmetry, 
             data=rgcm.example, adaptSteps=1000)

# model predictions based on 1000 samples from the posterior 
# distribution

preds <- predict(model)
summary(preds)
plot(preds)

# categorisation probabilities & model likelihoods based 
# on 1000 samples from the posterior distribution

preds <- predict(model, type="probabilities")
summary(preds)
plot(preds)

# model predictions based on mean of posterior samples

preds <- predict(model, sample.type="mean")
preds$predictions
summary(preds)

# model prediction based on user samples (vector)

usample <- c(1, 0.5, 0.5, 20, 1, 1, 2, 0, 0.6, 0.4, 0)
names(usample) <- colnames(model$posterior.samples[[1]])
preds <- predict(model, user.sample=usample)

# model prediction based on user samples (matrix)

usample <- matrix(c(c(1, 0.5, 0.5, 20, 1, 1, 2, 0, 0.6, 0.4, 0), 
                    c(1, 0.5, 0.5, 20, 1, 1, 2, 0, 0.1, 0.9, 0),
                    c(1, 0.5, 0.5, 20, 1, 1, 2, 0, 0.9, 0.1, 0)), 
                  byrow=T, nrow=3)
colnames(usample) <- colnames(model$posterior.samples[[1]])
preds <- predict(model, user.sample=usample, sample.type="user")
mean(preds$predictions[[1]]$outcome==preds$correct)
mean(preds$predictions[[2]]$outcome==preds$correct)

}
