\name{rgcm-package}
\alias{rgcm-package}
\alias{rgcm}
\docType{package}
\title{
Bayesian parameter estimation & prediction for the Generalised Context Model
}
\description{
Performs Bayesian parameter estimation and prediction for the Generalised Context Model with categorical predictors
}
\details{
\tabular{ll}{
Package: \tab rgcm\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-02-03\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most important functions %~~
}
\author{
M치rton S칩skuthy

Maintainer: M치rton S칩skuthy <marton.soskuthy@york.ac.uk>
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
